Move session logs to private
